Feature buy

+++ src/main/java/com/example/springbootproject/auth/dto/response/UserInfoResponse.java
@@ -1,5 +1,5 @@
 package com.example.springbootproject.auth.dto.response;
 
-public record UserInfoResponse(String name, String address, String email, Integer point) {
+public record UserInfoResponse(String name, String address, String email, Long point) {
 
 }
