+++ src/main/java/com/example/springbootproject/pointHistory/domain/PointHistory.java
@@ -23,7 +23,7 @@ public class PointHistory {
     private Long id;
 
     @Column(name = "POINTHISTORY_BALANCE")
-    private Integer balance;
+    private Long balance;
 
     @Column(name = "POINTHISTORY_TRANSACTION_TYPE")
     private Boolean transactionType; //T면 입금 F면 출금
@@ -32,13 +32,13 @@ public class PointHistory {
     private LocalDateTime createdAt;
 
     @Column(name = "PRICE")
-    private Integer price;
+    private Long price;
 
     @Column(name = "PRODUCTNAME")
     private String productName;
 
     @Column(name = "TRANSACTION_VOLUME")
-    private Integer transactionVolume;
+    private Long transactionVolume;
 
     @Column(name = "SIZE")
     private String size;
@@ -47,7 +47,7 @@ public class PointHistory {
     @ManyToOne
     private User user;
 
-    public void setBalance(Integer balance) {
+    public void setBalance(Long balance) {
         this.balance = balance;
     }
